use the same type (int)

+++ lib/ch01/binary_search.hpp
@@ -13,7 +13,7 @@ namespace binary_search
         while (lo <= hi)
         {
             // Key is in a[lo..hi] or not present.
-            size_t mid = lo + (hi - lo) / 2;
+            int mid = lo + (hi - lo) / 2;
             if (key < a[mid])
                 hi = mid - 1;
             else if (key > a[mid])
